move more timedelta constructor tests
